Deviations split
