+++ frontend/src/views/desktop/DeviationLogsView.vue
@@ -1,45 +1,16 @@
 <script setup lang="ts">
-import Badge from '@/components/desktop/shared/Badge.vue'
-import DesktopButton from '@/components/desktop/shared/DesktopButton.vue'
+import DeviationLogsFilter from '@/components/desktop/deviationlogs/DeviationLogsFilter.vue'
+import DeviationReviewCard from '@/components/desktop/deviationlogs/DeviationReviewCard.vue'
+import type {
+  DeviationLog,
+  DeviationReviewStatus,
+  DeviationUser,
+  UpdateDeviationPayload,
+} from '@/components/desktop/deviationlogs/types'
 import Loading from '@/components/desktop/shared/Loading.vue'
-import { delay } from '@/utils'
-import { Check, Eye, Save, X } from '@lucide/vue'
-import { computed, onMounted, ref } from 'vue'
 import Paginator from '@/components/desktop/shared/Paginator.vue'
-
-type DeviationCategory = 'IK_MAT' | 'IK_ALKOHOL' | 'OTHER'
-type DeviationReviewStatus = 'OPEN' | 'CLOSED'
-type OrgAccessLevel = 'OWNER' | 'MANAGER' | 'WORKER'
-
-type DeviationUser = {
-  id: number
-  legalName: string
-  accessLevel: OrgAccessLevel
-}
-
-type DeviationLog = {
-  id: number
-  ccpRecordId: number | null
-  routineRecordId: number | null
-  category: DeviationCategory
-  reportedBy: DeviationUser
-  reviewStatus: DeviationReviewStatus
-  reviewedBy: DeviationUser | null
-  reviewedAt: string | null
-  whatWentWrong: string
-  immediateActionTaken: string
-  potentialCause: string
-  potentialPreventativeMeasure: string
-  preventativeMeasureActuallyTaken: string
-  deviationReceivers: DeviationUser[]
-  createdAt: string
-}
-
-type UpdateDeviationPayload = {
-  reviewed_by: number
-  review_status: 'CLOSED'
-  preventative_measure_actually_taken: string
-}
+import { delay } from '@/utils'
+import { computed, onMounted, ref } from 'vue'
 
 const reviewFilter = ref<DeviationReviewStatus | 'ALL'>('OPEN')
 const loading = ref(true)
@@ -153,40 +124,6 @@ function shouldIncludeForCurrentUser(deviation: DeviationLog) {
   return isCurrentUserManagerOrOwner()
 }
 
-function getCategoryLabel(category: DeviationCategory) {
-  if (category === 'IK_MAT') {
-    return 'IK Mat'
-  }
-  if (category === 'IK_ALKOHOL') {
-    return 'IK Alkohol'
-  }
-
-  return 'Annet'
-}
-
-function getCategoryColor(category: DeviationCategory): 'navy' | 'cherry' {
-  return category === 'IK_ALKOHOL' ? 'cherry' : 'navy'
-}
-
-function getLinkedRecordText(deviation: DeviationLog) {
-  if (deviation.ccpRecordId !== null) {
-    return `CCP logg #${deviation.ccpRecordId}`
-  }
-  if (deviation.routineRecordId !== null) {
-    return `Rutine logg #${deviation.routineRecordId}`
-  }
-
-  return 'Ikke knyttet til en spesifikk logg'
-}
-
-function getReceiversText(deviation: DeviationLog) {
-  if (deviation.deviationReceivers.length === 0) {
-    return 'Ingen eksplisitte avviksmottakere - ledere og eier mottar avviket'
-  }
-
-  return deviation.deviationReceivers.map((user) => user.legalName).join(', ')
-}
-
 function formatDateTime(value: string | null) {
   if (!value) {
     return '-'
@@ -253,8 +190,13 @@ async function reviewDeviation(deviationId: number, payload: UpdateDeviationPayl
   }
 }
 
-function startReviewingDeviation(deviation: DeviationLog) {
-  if (submittingReview.value || deviation.reviewStatus === 'CLOSED') {
+function startReviewingDeviation(deviationId: number) {
+  if (submittingReview.value) {
+    return
+  }
+
+  const deviation = resource.value.find((entry) => entry.id === deviationId)
+  if (!deviation || deviation.reviewStatus === 'CLOSED') {
     return
   }
 
@@ -327,28 +269,10 @@ onMounted(async () => {
         <h1 class="instrument-serif-regular no-margin">Gjennomga mottatte avvik</h1>
         <hr class="navy-hr" />
 
-        <div class="filter-row">
-          <DesktopButton
-            content="Apne"
-            :icon="Check"
-            :on-click="() => (reviewFilter = 'OPEN')"
-            :disabled="reviewFilter === 'OPEN'"
-          />
-          <DesktopButton
-            content="Lukkede"
-            :icon="X"
-            :on-click="() => (reviewFilter = 'CLOSED')"
-            :disabled="reviewFilter === 'CLOSED'"
-            button-color="navy"
-          />
-          <DesktopButton
-            content="Alle"
-            :icon="Eye"
-            :on-click="() => (reviewFilter = 'ALL')"
-            :disabled="reviewFilter === 'ALL'"
-            button-color="navy"
-          />
-        </div>
+        <DeviationLogsFilter
+          :review-filter="reviewFilter"
+          @set-filter="(filter) => (reviewFilter = filter)"
+        />
 
         <Loading v-if="loading" />
         <p v-else-if="fetchError" class="error-message">Klarte ikke a hente avvik.</p>
@@ -358,111 +282,20 @@ onMounted(async () => {
             Ingen avvik i valgt visning.
           </p>
 
-          <article
+          <DeviationReviewCard
             v-for="deviation in filteredResource"
             :key="deviation.id"
-            class="deviation-card"
-            :class="deviation.reviewStatus === 'CLOSED' ? 'deviation-card-closed' : ''"
-          >
-            <div class="deviation-header-row">
-              <div class="deviation-header-meta">
-                <Badge :badge-color="getCategoryColor(deviation.category)">
-                  {{ getCategoryLabel(deviation.category) }}
-                </Badge>
-                <span class="deviation-id">#{{ deviation.id }}</span>
-                <span class="deviation-status">{{
-                  deviation.reviewStatus === 'OPEN' ? 'Apent' : 'Lukket'
-                }}</span>
-              </div>
-              <span class="deviation-date"
-                >Opprettet {{ formatDateTime(deviation.createdAt) }}</span
-              >
-            </div>
-
-            <div class="details-grid">
-              <div class="detail-box">
-                <span class="navy-subtitle">Rapportert av</span>
-                <span>{{ deviation.reportedBy.legalName }}</span>
-              </div>
-              <div class="detail-box">
-                <span class="navy-subtitle">Kobling</span>
-                <span>{{ getLinkedRecordText(deviation) }}</span>
-              </div>
-              <div class="detail-box detail-box-wide">
-                <span class="navy-subtitle">Avviksmottakere</span>
-                <span>{{ getReceiversText(deviation) }}</span>
-              </div>
-            </div>
-
-            <div class="deviation-section">
-              <span class="navy-subtitle">Hva gikk galt</span>
-              <p class="no-margin">{{ deviation.whatWentWrong }}</p>
-            </div>
-
-            <div class="deviation-section">
-              <span class="navy-subtitle">Umiddelbar handling</span>
-              <p class="no-margin">{{ deviation.immediateActionTaken }}</p>
-            </div>
-
-            <div class="deviation-section">
-              <span class="navy-subtitle">Potensiell arsak</span>
-              <p class="no-margin">{{ deviation.potentialCause }}</p>
-            </div>
-
-            <div class="deviation-section">
-              <span class="navy-subtitle">Potensielt forebyggende tiltak</span>
-              <p class="no-margin">{{ deviation.potentialPreventativeMeasure }}</p>
-            </div>
-
-            <div v-if="deviation.reviewStatus === 'CLOSED'" class="deviation-section resolved-box">
-              <span class="navy-subtitle">Forebyggende tiltak som faktisk ble gjort</span>
-              <p class="no-margin">{{ deviation.preventativeMeasureActuallyTaken }}</p>
-              <p class="resolved-meta no-margin">
-                Gjennomgatt av {{ deviation.reviewedBy?.legalName || '-' }}
-                {{ formatDateTime(deviation.reviewedAt) }}
-              </p>
-            </div>
-
-            <div
-              v-else-if="reviewingDeviationId === deviation.id"
-              class="deviation-section review-box"
-            >
-              <span class="navy-subtitle">Forebyggende tiltak som faktisk ble gjort</span>
-              <textarea
-                v-model="reviewText"
-                class="simple-text-input"
-                rows="4"
-                :disabled="submittingReview"
-                placeholder="Beskriv tiltaket som faktisk ble gjennomfort"
-              />
-              <div class="review-actions">
-                <DesktopButton
-                  :icon="Save"
-                  content="Lukk avvik"
-                  :on-click="() => submitDeviationReview(deviation.id)"
-                  :is-loading="submittingReview"
-                  loading-text="Lukker"
-                  :disabled="reviewText.trim().length === 0"
-                />
-                <DesktopButton
-                  :icon="X"
-                  content="Avbryt"
-                  button-color="cherry"
-                  :on-click="cancelReviewingDeviation"
-                  :disabled="submittingReview"
-                />
-              </div>
-              <p v-if="updateError" class="error-message">Klarte ikke a oppdatere avvik.</p>
-            </div>
-
-            <div v-else class="card-footer">
-              <DesktopButton
-                :icon="Check"
-                content="Gjennomga avvik"
-                :on-click="() => startReviewingDeviation(deviation)"
-              />
-            </div>
-          </article>
+            :deviation="deviation"
+            :is-reviewing="reviewingDeviationId === deviation.id"
+            :review-text="reviewingDeviationId === deviation.id ? reviewText : ''"
+            :submitting-review="submittingReview"
+            :update-error="reviewingDeviationId === deviation.id && updateError"
+            :format-date-time="formatDateTime"
+            @start-review="startReviewingDeviation"
+            @cancel-review="cancelReviewingDeviation"
+            @submit-review="submitDeviationReview"
+            @update-review-text="(value) => (reviewText = value)"
+          />
         </template>
       </div>
     </main>
@@ -493,96 +326,6 @@ main {
   width: calc(3 / 5 * 100%);
 }
 
-.filter-row {
-  display: flex;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-}
-
-.deviation-card {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-  border-radius: 1rem;
-  border: 1px solid var(--blue-navy-40);
-  background-color: var(--white-greek);
-  padding: 1rem;
-}
-
-.deviation-card-closed {
-  background-color: var(--blue-light-20);
-}
-
-.deviation-header-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0.75rem;
-  padding-bottom: 0.75rem;
-  border-bottom: 1px solid var(--blue-navy-20);
-}
-
-.deviation-header-meta {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-.deviation-id,
-.deviation-status,
-.deviation-date {
-  color: var(--blue-navy-80);
-}
-
-.details-grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 0.75rem;
-}
-
-.detail-box {
-  display: flex;
-  flex-direction: column;
-  gap: 0.2rem;
-  background-color: var(--blue-light-20);
-  border: 1px solid var(--blue-navy-20);
-  border-radius: 0.5rem;
-  padding: 0.6rem;
-}
-
-.detail-box-wide {
-  grid-column: span 2;
-}
-
-.deviation-section {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.resolved-box,
-.review-box {
-  border: 1px solid var(--blue-navy-20);
-  border-radius: 0.5rem;
-  background-color: var(--blue-light-20);
-  padding: 0.75rem;
-}
-
-.resolved-meta {
-  color: var(--blue-navy-80);
-}
-
-.review-actions {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.card-footer {
-  display: flex;
-  justify-content: flex-end;
-}
-
 .error-message {
   color: #b42318;
   margin: 0;
@@ -602,23 +345,6 @@ main {
 @media (max-width: 768px) {
   .deviation-logs-page-container {
     width: 100%;
-  }
-
-  .deviation-header-row {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .details-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .detail-box-wide {
-    grid-column: span 1;
-  }
-
-  .review-actions {
-    flex-direction: column;
   }
 }
 </style>
